refactor: move to _utils and rename data to event_args
