+++ src/ape/utils.py
@@ -10,29 +10,14 @@ from collections import namedtuple
 from functools import lru_cache
 from io import BytesIO
 from pathlib import Path
-from typing import (
-    TYPE_CHECKING,
-    Any,
-    ClassVar,
-    Dict,
-    Iterator,
-    List,
-    Mapping,
-    Optional,
-    Set,
-    Union,
-    cast,
-)
+from typing import TYPE_CHECKING, Any, ClassVar, Dict, List, Mapping, Optional, Set, cast
 
 import pygit2  # type: ignore
 import requests
 import yaml
-from eth_abi import grammar
 from eth_account import Account
 from eth_account.hdaccount import HDPath, seed_from_mnemonic
 from eth_utils import to_checksum_address as to_address
-from eth_utils.abi import collapse_if_tuple
-from ethpm_types.abi import EventABI, EventABIType
 from github import Github, UnknownObjectException
 from github.GitRelease import GitRelease
 from github.Organization import Organization
@@ -744,33 +729,6 @@ class BaseInterfaceModel(BaseInterface, BaseModel):
             kwargs["exclude_none"] = True
 
         return super().json(*args, **kwargs)
-
-
-class LogInputABICollection:
-    def __init__(self, abi: EventABI, values: List[EventABIType]):
-        self.abi = abi
-        self.values = values
-
-    @property
-    def names(self) -> List[str]:
-        return [abi.name for abi in self.values if abi.name]
-
-    @property
-    def normalized_values(self) -> List[Dict]:
-        return [abi.dict() for abi in self.values]
-
-    @property
-    def types(self) -> List[Union[str, Dict]]:
-        return [t for t in _get_event_abi_types(self.normalized_values)]
-
-
-def _get_event_abi_types(abi_inputs: List[Dict]) -> Iterator[Union[str, Dict]]:
-    for abi_input in abi_inputs:
-        abi_type = grammar.parse(abi_input["type"])
-        if abi_type.is_dynamic:
-            yield "bytes32"
-        else:
-            yield collapse_if_tuple(abi_input)
 
 
 __all__ = [
